pricepulse/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
beautifulsoup4==4.12.2
requests==2.31.0
apscheduler==3.10.1
lxml==4.9.3
python-dotenv==1.0.0
flask-sqlalchemy==3.0.5
pytz==2023.3
selenium==4.11.2
webdriver-manager==4.0.0